+++ src/FluxFlow.Components.Secrets/Contracts/SecretDiagnosticSeverity.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Components.Secrets.Contracts;
+
+public enum SecretDiagnosticSeverity
+{
+    Information = 1,
+    Warning = 2,
+    Error = 3
+}